ArmstrongNumbers.java
Given an integer. Check if the number is a Narcissistic number (Armstrong number) or not.
a. Example: Let us assume the number is 371. To check if the number is an Armstrong number, we should raise each digit to the power of 3 (number of the its digits). 3^3 + 7^3 + 1^3 = 371. If the sum equals to the number itself, it is considered to be an Armstrong number.
b. If user enters two integers, can you print out the list of all such numbers in a given range?