DiceMan.java
Write a program to simulate a game, “Dice Man”. The user (player) is asked to suggest a number that can be obtained by throwing two dice (a 6-sided fair dice) and then computer “throws two dice”. If the result is equal to what the user suggested, player wins. Otherwise, computer wins. The game is played in rounds:
a. program is terminated if user enters 0
b. prints an error message if the user enters a value which is impossible to obtain with two
dice.