GuessGame.java
Write a program to simulate the “Guess Game”. Your program takes a
positive random integer number not greater than A (to be entered by the user). It then lets user to guess the number. If the number is guessed correctly, user wins. Otherwise, the program helps user by indicating whether the number entered is LOWER or HIGHER.